lots of cleanup, more helper functions
